Extract the view from the base App function
